Moved moving burnin detectors to ocv
